requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
Werkzeug==1.0.1
dominate==2.6.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.1.0
importlib-metadata==4.0.1
iso8601==0.1.14
pyserial==3.5
PyYAML==5.4.1
readchar==3.0.4
serial==0.0.97
SQLAlchemy==1.4.14
typing-extensions==3.10.0.0
visitor==0.1.3
WTForms==2.3.3
zipp==3.4.1
PyMySQL==0.7.11
RPi.GPIO==0.7.0
paramiko==2.7.2
wakeonlan==2.0.1

